requirements.txt
Flask==2.2.3
mutagen==1.46.0
PyYAML==6.0
rich==13.3.2
SQLAlchemy==2.0.5.post1
